sre-agent/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sre-agent"
version = "0.1.0"
authors = [
    {name = "SRE Team", email = "sre@example.com"},
]
description = "SRE Agent using LangGraph for alert processing and solution recommendations"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "langgraph>=0.0.17",
    "langchain>=0.0.267",
    "langchain-openai>=0.0.2",
    "code-rag",
    "fastapi>=0.103.1",
    "pydantic>=2.4.2",
    "uvicorn>=0.23.2",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.7.0",
    "isort>=5.12.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88